create our own js file and load it on product pages by calling it from product.js
product.js is mapped to product pages in app.js in the PageClasses object.

It looks like you can only load one js file per page type so you need to edit the relevant one, or create a new page type and assign the JS to that via the PageClasses object.